UI fixes and changes to support AWS instances with threshold checks
