+++ seyren-core/src/main/java/com/seyren/core/util/config/SeyrenConfig.java
@@ -97,8 +97,6 @@ public class SeyrenConfig {
     private final String scriptType;
     private final String scriptResourceUrls;
     private final String graphiteRefreshRate;
-    private final String awsAccessKey;
-    private final String awsSecretKey;
 
     public SeyrenConfig() {
 
@@ -202,9 +200,6 @@ public class SeyrenConfig {
         this.scriptType = configOrDefault("SCRIPT_TYPE", "python");
         this.scriptResourceUrls = configOrDefault("SCRIPT_RESOURCE_URLS", "ERROR: None Defined");
 
-        //Aws
-        this.awsAccessKey = configOrDefault("AWS_ACCESS_KEY","");
-        this.awsSecretKey = configOrDefault("AWS_SECRET_KEY","");
     }
 
     @PostConstruct
@@ -530,18 +525,6 @@ public class SeyrenConfig {
     @JsonIgnore
     public String getGraphiteRefreshRate() {
         return graphiteRefreshRate;
-    }
-
-    @JsonIgnore
-    public String getAwsAccessKey()
-    {
-        return awsAccessKey;
-    }
-
-    @JsonIgnore
-    public String getAwsSecretKey()
-    {
-        return awsSecretKey;
     }
 
     private static String configOrDefault(String propertyName, String defaultValue) {
