+++ seyren-web/src/main/webapp/html/modal-partial-subscription.html
@@ -36,6 +36,7 @@
                             <option value="TWILIO">TWILIO</option>
                             <option value="VICTOROPS">VictorOps</option>
                             <option value="SCRIPT">Script</option>
+                            <option value="AWS_UNHEALTHY_INSTANCE">AWS_ASG</option>
                         </select>
                     </div>
                 </div>
